Add source code to spec-template.
The existing behavior is to display "When: ----" or similar for the block text.
This remains when com.athaydes.spockframework.report.showCodeBlocks is false.
When it is true, the code associated with the block is pasted.

CSS is not added yet. A further improvement would be to remove the blocks
entirely when showCodeBlocks is false.